fixed some tests, but sitll need to process rest.

+++ src/neo_protocol/responses/neo_transaction_result.rs
@@ -25,7 +25,7 @@ pub struct TransactionResult {
 	pub block_hash: Option<H256>,
 	pub confirmations: Option<i32>,
 	#[serde(rename = "blocktime")]
-	pub block_time: Option<i32>,
+	pub block_time: Option<u64>,
 	#[serde(rename = "vmstate")]
 	pub vm_state: Option<NeoVMStateType>,
 }
